LSM: Full security module stacking
Allow any combination of existing security modules,
including those using secids and security marked networking.

The interfaces used by filesystems to maintain security
attributes:
        security_inode_setsecctx
        security_inode_getsecctx
        security_inode_notifysecctx
have been trained to keep a full set of attributes
using the "lsm1='data1',lsm2='data2'" format.

A prctl interface has been added to identify which
security module should be invoked when secids are
translated to secctx and back. If none is specified the
first module will be used. This eliminates the ambiguity
of what data will be seen in user-space at the cost of
requiring user-space code to be explicit about what it
wants to see.

A set of secids are used in most cases where a single
secid was used in the past. A single secid is used within
secuirty modules, whereas a secid for each compiled module
that uses secids is maintained elsewhere. The secmarks used
by netfilter requires a mechanism to map multiple secids into
a single u32. That is provided.

The netlabel system is enhanced to allow it to
detect that security modules are trying to set network
labels that are incompatible. If incompatibilty is detected
setting the networks label will fail. As SELinux and Smack use
the interfaces differently, a system with both enabled requires
significant configuration attention to use network labeling
effectively.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>