LSM: Create and manage the lsmblob data structure.
When more than one security module is exporting data to
audit and networking sub-systems a single 32 bit integer
is no longer sufficient to represent the data. Add a
structure to be used instead.

The lsmblob structure is currently an array of
u32 "secids". There is an entry for each of the
security modules built into the system, based on
their respective configuration options. There is
an addition entry for containing network "secmarks"
should that be needed. Each registered LSM that
has hooks using secids is assigned a slot in the
lsmblob.

Helper functions for initializing (lsmblob_init),
comparing (lsmblob_equal) and checking (lsmblob_is_set)
these structures.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>