LSM: Specify which LSM to display
Create a new entry "display" in /proc/.../attr for controlling
which LSM security information is displayed for a process.
The name of an active LSM that supplies hooks for human readable
data may be written to "display" to set the value. The name of
the LSM currently in use can be read from "display".
At this point there can only be one LSM capable of display
active.

This affects /proc/.../attr/current and SO_PEERSEC.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>