LSM: Ensure the correct LSM context releaser
Add a new lsmcontext data structure to hold all the information
about a "security context", including the string, its size and
which LSM allocated the string. The allocation information is
necessary because LSMs have different policies regarding the
lifecycle of these strings. SELinux allocates and destroys
them on each use, whereas Smack provides a pointer to an entry
in a list that never goes away.

Change the security_release_secctx() interface to use the
lsmcontext and call only the appropiate LSM hook. Change
the callers of security_release_secctx() to provide the
correct type of data, introducing scaffolding where required.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>