NET: Add SO_PEERCONTEXT for multiple LSMs
The getsockopt SO_PEERSEC provides the LSM based security
information for a single module, but for reasons of backward
compatibility cannot include the information for multiple
modules. A new option SO_PEERCONTEXT is added to report the
security "context" of multiple modules using a "compound"
(or, in early discussions "hideous") format:
	lsm1='<value1>[,lsm2='<value2>']...

This is expected to be used by system services, including dbus-daemon.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>