LSM: Specify which LSM to display
Create a new entry "display" in /proc/.../attr for controlling
which LSM security information is displayed for a process.
The name of an active LSM that supplies hooks for human readable
data may be written to "display" to set the value. The name of the
LSM currently in use can be read from "display". At this point there
can only be one LSM capable of display active. A helper function
lsm_task_display() is provided to get the display slot for a task_struct.

Setting the "display" requires that all security modules using
setprocattr hooks allow the action. Each security module is
responcible for defining its policy. SELinux, Smack and AppArmor
hooks have been modified to allow setting the display in all cases.
This is not expected to last.

Note that procfs restricts writes in the attr directory to 'self'.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>