LSM: Use lsmcontext in security_dentry_init_security
Change the security_dentry_init_security() interface to
fill an lsmcontext structure instead of a void * data area
and a length. The lone caller of this interface is NFS4,
which may make copies of the data using its own mechanisms.
A rework of the nfs4 code to use the lsmcontext properly
is a significant project. SELinux is handled correctly, and
is the only current user.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>