LSM: lsm_context in security_dentry_init_security
Replace the (secctx,seclen) pointer pair with a single lsm_context
pointer to allow return of the LSM identifier along with the context
and context length. This allows security_release_secctx() to know how
to release the context. Callers have been modified to use or save the
returned data from the new structure.

Special care is taken in the NFS code, which uses the same data structure
for its own copied labels as it does for the data which comes from
security_dentry_init_security().  In the case of copied labels the data
has to be freed, not released.

The scaffolding funtion lsmcontext_init() is no longer needed and is
removed.

Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>
Cc: ceph-devel@vger.kernel.org
Cc: linux-nfs@vger.kernel.org